requirements.txt
streamlit>=1.20.0
python-docx>=0.8.11
openai>=0.27.0
sentence_transformers==2.2.2
tqdm
pandas
PyYAML
python-dotenv